Add proper support for multline
This inherits equation/displaymath but turns \\ into no-op.

This is reasonable behaviour since multline is usually used when the
page isn't wide enough, but this is not a constraint for webpages since
we can scroll.

(in any case, the point at which a line becomes too long is different
from that in pdf, so we aren't getting correct behaviour anyway)